backend/fileprocessor/requirements.txt
pytesseract==0.3.13
pillow==11.0.0
pdf2image==1.17.0 
Spire.Doc==12.7.1 